+++ packages/sdk/src/chains/solana/SolanaActions.ts
@@ -49,6 +49,7 @@ function createSolanaCoreContext(config: LombardConfig): SolanaCoreContext {
   return {
     env: config.env,
     partner: new PartnerConfiguration(config.partner),
+    auth: config.auth,
     getAuthToken: config.getAuthToken,
     getProvider: async (key) => {
       const getter = getProviderGetter(config.providers, key);