+++ packages/sdk/src/chains/starknet/StarknetActions.ts
@@ -38,6 +38,7 @@ function createStarknetCoreContext(config: LombardConfig): StarknetCoreContext {
   return {
     env: config.env,
     partner: new PartnerConfiguration(config.partner),
+    auth: config.auth,
     getAuthToken: config.getAuthToken,
     getProvider: async (key) => {
       const getter = getProviderGetter(config.providers, key);