+++ packages/sdk/src/shared/context/ContextBuilder.ts
@@ -81,6 +81,7 @@ function createCoreContext(config: LombardConfig): CoreContext {
     partner: new PartnerConfiguration(config.partner),
     getProvider: createProviderResolver(config),
     logger: undefined,
+    auth: config.auth,
     getAuthToken: config.getAuthToken,
   };
 }